+++ src/Routes.hs
@@ -13,6 +13,7 @@ routes = [ ("/",               method GET welcomePage)
          , ("/users/add",      methods [GET, POST] handleUserAdd)
          , ("/users/edit/:id", methods [GET, POST] handleUserEdit)
          , ("/users/put",      method PUT handleUserPut)
+         , ("/users/delete/:id", method DELETE handleUserDelete)
          , ("static",          serveDirectory "static")]
 
 welcomePage :: AppHandler