+++ test/SiteSpec.hs
@@ -4,12 +4,12 @@ import Data.ByteString
 import Test.Hspec
 import Snap.Snaplet.Test as T
 import Snap.Test
-import Data.Map as M
+import Data.Map as M hiding(delete)
 import Data.Either.Unwrap
 import Snap.Internal.Http.Types
 import Site
 import User.Handlers
-import EnvRead
+import IoInteractions
 
 spec :: Spec
 spec =
@@ -23,7 +23,7 @@ spec =
     context "when requesting /users/add" $
       it "should respond with status 200" $ do
         ci <- loadConnInfoFromEnv
-        er <- T.runHandler Nothing (get "/users/add" M.empty) handleUsers $ appInit ci
+        er <- T.runHandler Nothing (get "/users/add" M.empty) handleUserAdd $ appInit ci
         let r = fromRight er
         rspStatus r `shouldBe` 200
     context "when requesting /users/edit/ without the id parameter" $
@@ -55,5 +55,13 @@ spec =
               "\"password\":\"ibarra\",\"gender\":\"Male\",\"income\":1250}"
             request = put "/users/put" "application/json" reqBody
         er <- T.runHandler Nothing request handleUserPut $ appInit ci
+        let r = fromRight er
+        rspStatus r `shouldBe` 200
+    context "when requesting /users/delete/:id" $
+      it "should respond with status 200" $ do
+        ci <- loadConnInfoFromEnv
+        er <- T.runHandler Nothing 
+          (delete "/users/delete/id" $ M.fromList [("id", ["500"])]) 
+          handleUserDelete $ appInit ci
         let r = fromRight er
         rspStatus r `shouldBe` 200